refactor(server): absorb session SQLite backend
