+++ crates/container-sandbox/src/lib.rs
@@ -1,7 +1,7 @@
 //! Container Sandbox — self-hosted container execution via Docker Engine REST API.
 //!
-//! Decision: Core crate (not integration) because container execution is infrastructure,
-//! not an external service. Same pattern as `crates/session-sqldb/`.
+//! Decision: Infrastructure crate (not integration) because container execution is
+//! deployment infrastructure, not an external service.
 //!
 //! Decision: Docker Engine REST API directly, no `docker` CLI binary dependency.
 //! Workers can be containerized themselves — talking to the Docker Engine API
