+++ crates/server/src/app_builder.rs
@@ -669,10 +669,9 @@ impl ServerAppBuilder {
             ),
         );
 
-        let sqldb_backend = Arc::new(everruns_session_sqldb::InMemorySqlDbBackend::new());
-        let sqldb_store: Arc<dyn everruns_platform::session_sqldb::SessionSqlDbStore> = Arc::new(
-            everruns_session_sqldb::InMemorySqlDbStore::new(sqldb_backend),
-        );
+        let sqldb_backend = Arc::new(crate::session_sqldb::InMemorySqlDbBackend::new());
+        let sqldb_store: Arc<dyn everruns_platform::session_sqldb::SessionSqlDbStore> =
+            Arc::new(crate::session_sqldb::InMemorySqlDbStore::new(sqldb_backend));
         tracing::info!("Session SQL database store initialized (in-memory)");
 
         // =====================================================================
