+++ crates/server/src/session_sqldb/error.rs
