--- crates/session-sqldb/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "everruns-session-sqldb"
-publish = false
-readme = "README.md"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-description = "Session-scoped SQLite databases backed by PostgreSQL page-level storage"
-
-[dependencies]
-# SQLite
-rusqlite = { version = "0.39", features = ["bundled", "serialize", "column_decltype", "hooks"] }
-
-# Core types
-everruns-core = { path = "../core" }
-everruns-provider.workspace = true
-everruns-platform = { path = "../platform" }  # SessionSqlDbStore contract (EVE-897)
-
-# Async
-tokio = { workspace = true, features = ["rt", "time", "sync"] }
-async-trait.workspace = true
-
-# Sync
-parking_lot.workspace = true
-
-# Serialization
-serde_json.workspace = true
-
-# Types
-chrono.workspace = true
-uuid.workspace = true
-
-# Error handling
-thiserror.workspace = true
-
-# Logging
-tracing.workspace = true
-
-# Base64 for BLOB encoding
-base64.workspace = true
-
-[dev-dependencies]
-tokio = { workspace = true, features = ["full", "test-util"] }
